Enum paige::Tag

pub enum Tag {
    A,
    Abbr,
    Acronym,
    Address,
    Applet,
    Area,
    Article,
    Aside,
    Audio,
    B,
    Base,
    Basefont,
    Bdi,
    Bdo,
    Big,
    Blockquote,
    Body,
    Br,
    Button,
    Canvas,
    Caption,
    Center,
    Cite,
    Code,
    Col,
    Colgroup,
    Data,
    Datalist,
    Dd,
    Del,
    Details,
    Dfn,
    Dialog,
    Dir,
    Div,
    Dl,
    Dt,
    Em,
    Embed,
    Fieldset,
    FigCaption,
    Figure,
    Font,
    Footer,
    Form,
    Frame,
    Frameset,
    H1,
    Head,
    Header,
    Hr,
    Html,
    I,
    IFrame,
    Img,
    Input,
    Ins,
    Kbd,
    Label,
    Legend,
    Li,
    Link,
    Main,
    Map,
    Mark,
    Meta,
    Meter,
    Nav,
    Noframes,
    Noscript,
    Object,
    Ol,
    Optgroup,
    Option_,
    Output,
    P,
    Param,
    Picture,
    Pre,
    Progress,
    Q,
    Rp,
    Rt,
    Ruby,
    S,
    Samp,
    Script,
    Section,
    Select,
    Small,
    Source,
    Span,
    Strike,
    Strong,
    Style,
    Sub,
    Summary,
    Sup,
    Svg,
    Table,
    Tbody,
    Td,
    Template,
    TextArea,
    Tfoot,
    Th,
    Thead,
    Time,
    Title,
    Tr,
    Track,
    Tt,
    U,
    Ul,
    Var,
    Video,
    Wbr,
}

HTML tags

Variants

A

Abbr

Acronym

Address

Applet

Area

Article

Aside

Audio

B

Base

Basefont

Bdi

Bdo

Big

Blockquote

Body

Br

Button

Canvas

Caption

Center

Cite

Code

Col

Colgroup

Data

Datalist

Dd

Del

Details

Dfn

Dialog

Dir

Div

Dl

Dt

Em

Embed

Fieldset

FigCaption

Figure

Font

Footer

Form

Frame

Frameset

H1

Head

Header

Hr

Html

I

IFrame

Img

Input

Ins

Kbd

Label

Legend

Li

Link

Main

Map

Mark

Meta

Meter

Nav

Noframes

Noscript

Object

Ol

Optgroup

Option_

Output

P

Param

Picture

Pre

Progress

Q

Rp

Rt

Ruby

S

Samp

Script

Section

Select

Small

Source

Span

Strike

Strong

Style

Sub

Summary

Sup

Svg

Table

Tbody

Td

Template

TextArea

Tfoot

Th

Thead

Time

Title

Tr

Track

Tt

U

Ul

Var

Video

Wbr

Methods

impl Tag

pub fn make_element(self) -> El

Consumes this Tag and returns an El.

pub fn attributes(self, attrs: &[Attr]) -> El

Consumes a Tag and returns an El with the specified attributes.

pub fn style(self, props: &[Prop]) -> El

Consumes a Tag and returns an El with the specified style properties.

pub fn content(self, content: &[El]) -> El

Consumes a Tag and returns an El with the specified content (children).